+++ pueblo/ngr/model.py
@@ -5,6 +5,7 @@ class ItemType(Enum):
     DOTNET = "dotnet"
     JAVA = "java"
     JAVASCRIPT = "javascript"
+    JULIA = "julia"
     MAKE = "make"
     PHP = "php"
     PYTHON = "python"