+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - ngr: Use Makefile runner as fallback
+- ngr: Add runner for Julia projects
 
 ## 2023-10-30 v0.0.1
 - Add `pueblo.util.logging.setup_logging`, nomen est omen